+++ cpp/tests/io/parquet_test.cpp
@@ -946,6 +946,68 @@ TEST_F(ParquetChunkedWriterTest, Strings)
   CUDF_TEST_EXPECT_TABLES_EQUAL(*result.tbl, *expected);
 }
 
+TEST_F(ParquetChunkedWriterTest, ListColumn)
+{
+  auto valids  = cudf::test::make_counting_transform_iterator(0, [](auto i) { return i % 2; });
+  auto valids2 = cudf::test::make_counting_transform_iterator(0, [](auto i) { return i != 3; });
+
+  using lcw = cudf::test::lists_column_wrapper<int32_t>;
+
+  // COL0 (Same nullability) ====================
+  // [NULL, 2, NULL]
+  // []
+  // [4, 5]
+  // NULL
+  lcw col0_tbl0{{{{1, 2, 3}, valids}, {}, {4, 5}, {}}, valids2};
+
+  // [7, 8, 9]
+  // []
+  // [NULL, 11]
+  // NULL
+  lcw col0_tbl1{{{7, 8, 9}, {}, {{10, 11}, valids}, {}}, valids2};
+
+  // COL1 (Nullability different in different chunks, test of merging nullability in writer)
+  // [NULL, 2, NULL]
+  // []
+  // [4, 5]
+  // []
+  lcw col1_tbl0{{{1, 2, 3}, valids}, {}, {4, 5}, {}};
+
+  // [7, 8, 9]
+  // []
+  // [10, 11]
+  // NULL
+  lcw col1_tbl1{{{7, 8, 9}, {}, {10, 11}, {}}, valids2};
+
+  // COL2 (non-nested columns to test proper schema construction)
+  size_t num_rows_tbl0 = static_cast<cudf::column_view>(col0_tbl0).size();
+  size_t num_rows_tbl1 = static_cast<cudf::column_view>(col0_tbl1).size();
+  auto seq_col0        = random_values<int>(num_rows_tbl0);
+  auto seq_col1        = random_values<int>(num_rows_tbl1);
+
+  column_wrapper<int> col2_tbl0{seq_col0.begin(), seq_col0.end(), valids};
+  column_wrapper<int> col2_tbl1{seq_col1.begin(), seq_col1.end(), valids2};
+
+  auto tbl0 = table_view({col0_tbl0, col1_tbl0, col2_tbl0});
+  auto tbl1 = table_view({col0_tbl1, col1_tbl1, col2_tbl1});
+
+  auto expected = cudf::concatenate({tbl0, tbl1});
+
+  auto filepath = temp_env->get_temp_filepath("ChunkedLists.parquet");
+  cudf_io::chunked_parquet_writer_options args =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath});
+  auto state = cudf_io::write_parquet_chunked_begin(args);
+  cudf_io::write_parquet_chunked(tbl0, state);
+  cudf_io::write_parquet_chunked(tbl1, state);
+  cudf_io::write_parquet_chunked_end(state);
+
+  cudf_io::parquet_reader_options read_opts =
+    cudf_io::parquet_reader_options::builder(cudf_io::source_info{filepath});
+  auto result = cudf_io::read_parquet(read_opts);
+
+  CUDF_TEST_EXPECT_TABLES_EQUAL(*result.tbl, *expected);
+}
+
 TEST_F(ParquetChunkedWriterTest, MismatchedTypes)
 {
   srand(31337);
@@ -974,6 +1036,185 @@ TEST_F(ParquetChunkedWriterTest, MismatchedStructure)
   cudf_io::write_parquet_chunked(*table1, state);
   EXPECT_THROW(cudf_io::write_parquet_chunked(*table2, state), cudf::logic_error);
   cudf_io::write_parquet_chunked_end(state);
+}
+
+TEST_F(ParquetChunkedWriterTest, MismatchedStructureList)
+{
+  auto valids  = cudf::test::make_counting_transform_iterator(0, [](auto i) { return i % 2; });
+  auto valids2 = cudf::test::make_counting_transform_iterator(0, [](auto i) { return i != 3; });
+
+  using lcw = cudf::test::lists_column_wrapper<int32_t>;
+
+  // COL0 (mismatched depth) ====================
+  // [NULL, 2, NULL]
+  // []
+  // [4, 5]
+  // NULL
+  lcw col00{{{{1, 2, 3}, valids}, {}, {4, 5}, {}}, valids2};
+
+  // [[1, 2, 3], [], [4, 5], [], [0, 6, 0]]
+  // [[7, 8]]
+  // []
+  // [[]]
+  lcw col01{{{1, 2, 3}, {}, {4, 5}, {}, {0, 6, 0}}, {{7, 8}}, lcw{}, lcw{lcw{}}};
+
+  // COL2 (non-nested columns to test proper schema construction)
+  size_t num_rows = static_cast<cudf::column_view>(col00).size();
+  auto seq_col0   = random_values<int>(num_rows);
+  auto seq_col1   = random_values<int>(num_rows);
+
+  column_wrapper<int> col10{seq_col0.begin(), seq_col0.end(), valids};
+  column_wrapper<int> col11{seq_col1.begin(), seq_col1.end(), valids2};
+
+  auto tbl0 = table_view({col00, col10});
+  auto tbl1 = table_view({col01, col11});
+
+  auto filepath = temp_env->get_temp_filepath("ChunkedLists.parquet");
+  cudf_io::chunked_parquet_writer_options args =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath});
+  auto state = cudf_io::write_parquet_chunked_begin(args);
+  cudf_io::write_parquet_chunked(tbl0, state);
+  CUDF_EXPECT_THROW_MESSAGE(cudf_io::write_parquet_chunked(tbl1, state),
+                            "Mismatch in schema between multiple calls to write_chunk");
+}
+
+TEST_F(ParquetChunkedWriterTest, DifferentNullability)
+{
+  srand(31337);
+  auto table1 = create_random_fixed_table<int>(5, 5, true);
+  auto table2 = create_random_fixed_table<int>(5, 5, false);
+
+  auto full_table = cudf::concatenate({*table1, *table2});
+
+  auto filepath = temp_env->get_temp_filepath("ChunkedNullable.parquet");
+  cudf_io::chunked_parquet_writer_options args =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath});
+  auto state = cudf_io::write_parquet_chunked_begin(args);
+  cudf_io::write_parquet_chunked(*table1, state);
+  cudf_io::write_parquet_chunked(*table2, state);
+  cudf_io::write_parquet_chunked_end(state);
+
+  cudf_io::parquet_reader_options read_opts =
+    cudf_io::parquet_reader_options::builder(cudf_io::source_info{filepath});
+  auto result = cudf_io::read_parquet(read_opts);
+
+  CUDF_TEST_EXPECT_TABLES_EQUAL(*result.tbl, *full_table);
+}
+
+TEST_F(ParquetChunkedWriterTest, ForcedNullability)
+{
+  srand(31337);
+  auto table1 = create_random_fixed_table<int>(5, 5, false);
+  auto table2 = create_random_fixed_table<int>(5, 5, false);
+
+  auto full_table = cudf::concatenate({*table1, *table2});
+
+  auto filepath = temp_env->get_temp_filepath("ChunkedNoNullable.parquet");
+
+  cudf::io::table_metadata_with_nullability nullable_metadata;
+
+  // In the absence of prescribed per-column nullability in metadata, the writer assumes the worst
+  // and considers all columns nullable. However cudf::concatenate will not force nulls in case no
+  // columns are nullable. To get the expected result, we tell the writer the nullability of all
+  // columns in advance.
+  nullable_metadata.column_nullable.insert(nullable_metadata.column_nullable.begin(), 5, false);
+
+  cudf_io::chunked_parquet_writer_options args =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath})
+      .nullable_metadata(&nullable_metadata);
+  auto state = cudf_io::write_parquet_chunked_begin(args);
+  cudf_io::write_parquet_chunked(*table1, state);
+  cudf_io::write_parquet_chunked(*table2, state);
+  cudf_io::write_parquet_chunked_end(state);
+
+  cudf_io::parquet_reader_options read_opts =
+    cudf_io::parquet_reader_options::builder(cudf_io::source_info{filepath});
+  auto result = cudf_io::read_parquet(read_opts);
+
+  CUDF_TEST_EXPECT_TABLES_EQUAL(*result.tbl, *full_table);
+}
+
+TEST_F(ParquetChunkedWriterTest, ForcedNullabilityList)
+{
+  srand(31337);
+
+  auto valids  = cudf::test::make_counting_transform_iterator(0, [](auto i) { return i % 2; });
+  auto valids2 = cudf::test::make_counting_transform_iterator(0, [](auto i) { return i != 3; });
+
+  using lcw = cudf::test::lists_column_wrapper<int32_t>;
+
+  cudf::io::table_metadata_with_nullability nullable_metadata;
+
+  // COL0 ====================
+  // [1, 2, 3]
+  // []
+  // [4, 5]
+  // NULL
+  lcw col00{{{1, 2, 3}, {}, {4, 5}, {}}, valids2};
+
+  // [7]
+  // []
+  // [8, 9, 10, 11]
+  // NULL
+  lcw col01{{{7}, {}, {8, 9, 10, 11}, {}}, valids2};
+
+  nullable_metadata.column_nullable.push_back(true);   // List is nullable at first (root) level
+  nullable_metadata.column_nullable.push_back(false);  // non-nullable at second (leaf) level
+
+  // COL1 (non-nested columns to test proper schema construction)
+  size_t num_rows = static_cast<cudf::column_view>(col00).size();
+  auto seq_col0   = random_values<int>(num_rows);
+  auto seq_col1   = random_values<int>(num_rows);
+
+  column_wrapper<int> col10{seq_col0.begin(), seq_col0.end(), valids};
+  column_wrapper<int> col11{seq_col1.begin(), seq_col1.end(), valids2};
+
+  nullable_metadata.column_nullable.push_back(true);
+
+  auto table1 = table_view({col00, col10});
+  auto table2 = table_view({col01, col11});
+
+  auto full_table = cudf::concatenate({table1, table2});
+
+  auto filepath = temp_env->get_temp_filepath("ChunkedListNullable.parquet");
+
+  cudf_io::chunked_parquet_writer_options args =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath})
+      .nullable_metadata(&nullable_metadata);
+  auto state = cudf_io::write_parquet_chunked_begin(args);
+  cudf_io::write_parquet_chunked(table1, state);
+  cudf_io::write_parquet_chunked(table2, state);
+  cudf_io::write_parquet_chunked_end(state);
+
+  cudf_io::parquet_reader_options read_opts =
+    cudf_io::parquet_reader_options::builder(cudf_io::source_info{filepath});
+  auto result = cudf_io::read_parquet(read_opts);
+
+  CUDF_TEST_EXPECT_TABLES_EQUAL(*result.tbl, *full_table);
+}
+
+TEST_F(ParquetChunkedWriterTest, WrongNullability)
+{
+  srand(31337);
+  auto table1 = create_random_fixed_table<int>(5, 5, false);
+
+  auto filepath = temp_env->get_temp_filepath("ChunkedWrongNullable.parquet");
+
+  cudf::io::table_metadata_with_nullability nullable_metadata;
+  nullable_metadata.column_nullable.insert(nullable_metadata.column_nullable.begin(), 6, true);
+  cudf_io::chunked_parquet_writer_options args =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath})
+      .nullable_metadata(&nullable_metadata);
+  auto state = cudf_io::write_parquet_chunked_begin(args);
+  EXPECT_THROW(cudf_io::write_parquet_chunked(*table1, state), cudf::logic_error);
+
+  nullable_metadata.column_nullable.clear();
+  nullable_metadata.column_nullable.insert(nullable_metadata.column_nullable.begin(), 4, true);
+  cudf_io::chunked_parquet_writer_options args2 =
+    cudf_io::chunked_parquet_writer_options::builder(cudf_io::sink_info{filepath})
+      .nullable_metadata(&nullable_metadata);
+  state = cudf_io::write_parquet_chunked_begin(args2);
+  EXPECT_THROW(cudf_io::write_parquet_chunked(*table1, state), cudf::logic_error);
 }
 
 TEST_F(ParquetChunkedWriterTest, ReadRowGroups)